feat : 메인 뷰 stackview 구조 개선 - MainCardInfoStackView 추가 #3
